tools/make/requirements.txt
cairo-lang==0.12.2
typeguard==2.13.3
protobuf==3.20.3
inquirer
python-dotenv
pysha3